restructure client directory
