+++ client/assets/pages/events_page.scss
