+++ client/components/SocialClub/ClubInfo.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React from 'react';
 
 function ClubInfo({ joinClub }) {
   return (
