+++ client/index.js
@@ -2,9 +2,12 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from 'react-redux';
 import { Router, browserHistory } from 'react-router';
+
 import configureStore from './store/configureStore';
 import routes from './routes';
-import './App/assets/style.scss';
+
+// styles
+import './assets/style.scss'
 
 const store = configureStore();
