+++ client/routes.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
-import App from './App/app';
 
-import EventPage from './App/components/events/event/Index';
-import HomePage from './App/components/HomePage';
-import SocialClubPage from './App/components/SocialClubPage';
+// components
+import App from './components/App';
+import EventPage from './components/Events/index';
+import HomePage from './components/HomePage/HomePage';
+import SocialClubPage from './components/SocialClub/SocialClubPage';
 
 export default
   <Route path="/" component={App}>
